#78 Rename TemplateAttributeType to CkAttributeArrayValue
